+++ paddle/fluid/framework/ir/pass.h
@@ -26,7 +26,7 @@ limitations under the License. */
 #include "paddle/fluid/framework/ir/node.h"
 #include "paddle/fluid/framework/program_desc.h"
 #include "paddle/fluid/platform/variant.h"
-#include "paddle/fluid/utils/any.hpp"
+#include "paddle/utils/any.h"
 
 namespace paddle {
 namespace framework {
