+++ paddle/fluid/inference/CMakeLists.txt
@@ -39,7 +39,7 @@ get_property(fluid_modules GLOBAL PROPERTY FLUID_MODULES)
 # Adapt to custom op mechanism: Include the header files related to the data type
 # to avoid exposing the path of the underlying file
 include_directories(${PADDLE_SOURCE_DIR}/paddle/fluid/platform)
-include_directories(${PADDLE_SOURCE_DIR}/paddle/fluid/utils)
+include_directories(${PADDLE_SOURCE_DIR}/paddle/utils)
 
 add_subdirectory(api)
